Moved LoaderProblem to int pkg
