+++ andhow/src/test/java/yarnandtail/andhow/load/JndiLoaderTest.java
@@ -1,5 +1,6 @@
 package yarnandtail.andhow.load;
 
+import yarnandtail.andhow.internal.LoaderProblem;
 import java.time.LocalDateTime;
 import java.util.List;
 import org.junit.Test;
